OriginTrail Mainnet Release v5.1.1

+++ modules/service/dataset-pruning-service.js
@@ -1,5 +1,9 @@
 const Models = require('../../models/index');
 const { QueryTypes } = require('sequelize');
+const sequelizeConfig = require('../../config/sequelizeConfig');
+const constants = require('./../constants');
+
+const defaultBackupFolderPath = '/ot-node/backup/';
 
 class DatasetPruningService {
     /**
@@ -9,6 +13,8 @@ class DatasetPruningService {
     constructor(ctx) {
         this.logger = ctx.logger;
         this.graphStorage = ctx.graphStorage;
+        this.diskService = ctx.diskService;
+        this.config = ctx.config;
     }
 
     getIdsForPruning(
@@ -77,7 +83,6 @@ class DatasetPruningService {
                 }
             }
         });
-        this.logger.trace(`Found ${datasetsToBeDeleted.length} datasets for pruning`);
 
         return {
             datasetsToBeDeleted,
@@ -110,11 +115,13 @@ class DatasetPruningService {
     }
 
     async removeEntriesWithId(table, idArray) {
-        await Models[table].destroy({
-            where: {
-                id: { [Models.sequelize.Op.in]: idArray },
-            },
-        });
+        if (idArray && idArray.length > 0) {
+            await Models[table].destroy({
+                where: {
+                    id: { [Models.sequelize.Op.in]: idArray },
+                },
+            });
+        }
     }
 
     async removeDatasetsFromGraphDb(datasets) {
@@ -208,6 +215,147 @@ class DatasetPruningService {
             }
         });
         return repackedDatasets;
+    }
+
+    shouldPruneLowEstimatedValueDatasets() {
+        try {
+            if (!this.config.dataset_pruning.low_estimated_value_datasets.enabled) {
+                return false;
+            }
+            const diskSpace = this.diskService.getRootFolderInfo();
+            const freeSpacePercentage = (100 * diskSpace.free) / diskSpace.total;
+            const arangoDbEngineFolderSize = this.diskService
+                .getFolderSize(this.config.database.engine_folder_path);
+            const minimumArangoDbFolderSizeForPruning = 0.2 * diskSpace.total;
+            if (freeSpacePercentage > this.config.dataset_pruning
+                .low_estimated_value_datasets.minimum_free_space_percentage) {
+                this.logger.debug(`There is enough free space on the disk, low estimated value datasets pruning will not be executed. Disk free space: ${freeSpacePercentage}%`);
+                return false;
+            }
+
+            if (this.diskService.folderExists(defaultBackupFolderPath)
+                && this.diskService.getFolderSize(defaultBackupFolderPath) > 0) {
+                this.logger.warn('Detected ot-node backup on machine. please remove backup data in order to enable low estimated value datasets pruning!');
+                return false;
+            }
+
+            if (arangoDbEngineFolderSize < minimumArangoDbFolderSizeForPruning) {
+                this.logger.warn('Reached minimum Graph DB folder size, low estimated value datasets wont be pruned. ' +
+                    `Minimum size of Graph DB is 20% of total disk size. Current Graph DB folder size is: ${arangoDbEngineFolderSize}kb`);
+                return false;
+            }
+            return true;
+        } catch (error) {
+            this.logger.error('Error while trying to determine should low estimated datasets be pruned. Error: ', error.message);
+            return false;
+        }
+    }
+
+    async findLowEstimatedValueDatasets() {
+        const queryString = 'select di.id as data_info_id, di.data_set_id, bid.status as bid_status, ' +
+            'bid.id as bid_id, (di.import_timestamp + bid.holding_time_in_minutes*60000) as expiry,' +
+            'offer.id as offer_id, pd.id as purchase_id, di.import_timestamp ' +
+            'from data_info as di ' +
+            'left join offers as offer on di.data_set_id = offer.data_set_id ' +
+            'inner join bids as bid on di.data_set_id = bid.data_set_id ' +
+            'left join purchased_data as pd on di.data_set_id = pd.data_set_id ' +
+            'order by (di.import_timestamp + bid.holding_time_in_minutes*60000)';
+        return Models.sequelize.query(queryString, { type: QueryTypes.SELECT });
+    }
+
+    /**
+     * Returns repacked dataset information in format:
+     * {
+     *     [data_set_id]: {
+     *         dataInfos: [],
+     *         bids: [],
+     *         expiry: timestamp,
+     *         chosen: false,
+     *         foundOffer: false,
+     *         foundPurchase: false,
+     *     }
+     * }
+     * @param datasets
+     * @returns {{}}
+     */
+    repackLowEstimatedValueDatasets(datasets) {
+        const repackedDatasets = {};
+        datasets.forEach((dataset) => {
+            if (!repackedDatasets[dataset.data_set_id]) {
+                repackedDatasets[dataset.data_set_id] = {
+                    dataInfos: [],
+                    bids: [],
+                };
+            }
+            if (!repackedDatasets[dataset.data_set_id].expiry
+                || repackedDatasets[dataset.data_set_id].expiry < dataset.expiry) {
+                repackedDatasets[dataset.data_set_id].expiry = dataset.expiry;
+            }
+            const foundDataInfoId = repackedDatasets[dataset.data_set_id].dataInfos
+                .includes(dataset.data_info_id);
+            if (!foundDataInfoId) {
+                repackedDatasets[dataset.data_set_id].dataInfos.push(dataset.data_info_id);
+                repackedDatasets[dataset.data_set_id].importTimestamp = dataset.import_timestamp;
+            }
+
+            if (dataset.bid_id) {
+                const foundBidId = repackedDatasets[dataset.data_set_id].bids
+                    .includes(dataset.bid_id);
+                if (!foundBidId) {
+                    repackedDatasets[dataset.data_set_id].bids.push(dataset.bid_id);
+                    if (dataset.bid_status === 'CHOSEN') {
+                        repackedDatasets[dataset.data_set_id].chosen = true;
+                    }
+                }
+            }
+            if (dataset.offer_id) {
+                repackedDatasets[dataset.data_set_id].foundOffer = true;
+            }
+            if (dataset.purchase_id) {
+                repackedDatasets[dataset.data_set_id].foundPurchase = true;
+            }
+        });
+        return repackedDatasets;
+    }
+
+    getLowEstimatedValueIdsForPruning(repackedDatasets) {
+        const idsForPruning = {
+            datasetsToBeDeleted: [],
+            dataInfoIdToBeDeleted: [],
+            bidIdToBeDeleted: [],
+        };
+
+        let datasetsArray = [];
+        Object.keys(repackedDatasets).forEach((key) => {
+            const dataset = repackedDatasets[key];
+            if (!dataset.foundOffer && !dataset.foundPurchase && !dataset.chosen) {
+                datasetsArray.push({
+                    expiry: dataset.expiry,
+                    datasetId: {
+                        id: key,
+                        importTimestamp: dataset.importTimestamp,
+                    },
+                    dataInfoIds: dataset.dataInfos,
+                    bidsIds: dataset.bids,
+                });
+            }
+        });
+
+        datasetsArray = datasetsArray
+            .sort((a, b) => a.expiry - b.expiry)
+            .splice(0, constants.LOW_ESTIMATED_VALUE_DATASETS_PRUNING_BATCH_NUMBER);
+
+        datasetsArray.forEach((dataset) => {
+            idsForPruning.datasetsToBeDeleted.push({
+                datasetId: dataset.datasetId.id,
+                importTimestamp: dataset.datasetId.importTimestamp,
+            });
+            idsForPruning.dataInfoIdToBeDeleted =
+                idsForPruning.dataInfoIdToBeDeleted.concat(dataset.dataInfoIds);
+            idsForPruning.bidIdToBeDeleted =
+                idsForPruning.bidIdToBeDeleted.concat(dataset.bidsIds);
+        });
+        return idsForPruning;
     }
 }
